+++ react-app/src/hooks/api/useBCAssessmentApi.ts
@@ -33,14 +33,14 @@ const useBCAssessmentApi = () => {
   const url = window.location.href.includes('pims.gov.bc.ca')
     ? 'https://apps.gov.bc.ca/ext/sgw/geo.bca?REQUEST=GetFeature&SERVICE=WFS&VERSION=2.0.0&typeName=geo.bca:WHSE_HUMAN_CULTURAL_ECONOMIC.BCA_FOLIO_GNRL_PROP_VALUES_SV&outputFormat=application/json'
     : 'https://test.apps.gov.bc.ca/ext/sgw/geo.bca?REQUEST=GetFeature&SERVICE=WFS&VERSION=2.0.0&typeName=geo.bca:WHSE_HUMAN_CULTURAL_ECONOMIC.BCA_FOLIO_GNRL_PROP_VALUES_SV&outputFormat=application/json';
-  // https://test.apps.gov.bc.ca/ext/sgw/geo.bca?REQUEST=GetFeature&SERVICE=WFS&VERSION=2.0.0&typeName=geo.bca:WHSE_HUMAN_CULTURAL_ECONOMIC.BCA_FOLIO_GNRL_PROP_VALUES_SV&outputFormat=application/json&srsName=EPSG:4326&CQL_FILTER=OBJECTID=894817399
+
   const getBCAssessmentByLocation = async (lng: number, lat: number) => {
     const finalUrl = `${url}&srsName=EPSG:4326&CQL_FILTER=CONTAINS(SHAPE,SRID=4326;POINT(${lng} ${lat}))`;
     const response = await fetch(finalUrl, {
       credentials: 'include',
     });
+    if (!response.ok) return;
     const body = await response.json();
-    console.log(body);
     return body as FeatureCollection<Geometry, BCAssessmentProperties>;
   };
